Promote LTSA call to parent component
