+++ frontend/src/features/mapSideBar/components/tabs/TitleOwnership.tsx
@@ -2,32 +2,14 @@ import { ILTSAOrderModel } from 'actions/parcelsActions';
 import { Charges } from 'components/ltsa/Charges';
 import { OwnershipDetails } from 'components/ltsa/OwnershipInformation';
 import { TitleDetails } from 'components/ltsa/TitleDetails';
-import { getIn, useFormikContext } from 'formik';
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 
 interface ITitleOwnershipProps {
-  withNameSpace: Function;
+  ltsa: ILTSAOrderModel;
 }
 
 export const TitleOwnership: React.FC<any> = (props: ITitleOwnershipProps) => {
-  const { withNameSpace } = props;
-  const formikProps = useFormikContext();
-  const [ltsa, setLTSA] = useState<ILTSAOrderModel | undefined>(undefined);
-
-  useEffect(() => {
-    getLTSAInfo();
-  }, [formikProps]);
-
-  const getLTSAInfo = useCallback(async () => {
-    const ltsaInfo: ILTSAOrderModel | undefined = await getIn(
-      formikProps.values,
-      withNameSpace('ltsa'),
-    );
-    if (ltsaInfo) {
-      setLTSA(ltsaInfo);
-    }
-  }, [formikProps]);
-
+  const { ltsa } = props;
   return ltsa ? (
     <>
       <p
